relief-sync/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==0.28
python-multipart
pydantic
graphviz==0.20.3
networkx
xmltodict
python-dotenv
mysql-connector-python